The Travelling Salesman Problem (TSP)
TSP is a classic algorithmic problem in the field of computer science
and operations research, focusing on optimization.
It seeks the shortest possible route that visits every point in a set
of locations just once.
OR-Tools is open source software for combinatorial optimization,
which seeks to find the best solution to a problem out of a very l
arge set of possible solutions.